ATM Banking System project using Python and MySQL.
This ATM Banking System project is developed using Python as the front-end and MySQL as the back-end database. The system simulates basic banking operations, allowing users to perform essential tasks such as account creation, balance inquiry, cash withdrawal, and deposit transactions. It aims to provide a user-friendly interface for managing banking operations with a focus on security and efficiency.